+++ resources/views/layouts/inc/content-messages.blade.php
@@ -1,4 +1,8 @@
 <div class="content-message">
-    <a href="#" style="color: rgba(255, 255, 255, 0.9); display: inline-block; margin-right: 10px; text-decoration: none;">Внимание! Это просто пример сообщения, которое можно отключить в шаблоне!</a>
-    <a href="#" class="btn btn-default btn-sm" style="margin-top: -5px; border: 0px; box-shadow: none; color: rgb(243, 156, 18); font-weight: 600; background: rgb(255, 255, 255);">Показать больше</a>
+    <a href="#" style="color: rgba(255, 255, 255, 0.9); display: inline-block; margin-right: 10px; text-decoration: none;">Attention! This is just an example of a message that can be disabled in the template [lte::layouts.inc.content-messages]</a>
+    <a href="#"
+       class="btn btn-default btn-sm"
+       style="margin-top: -5px; border: 0px; box-shadow: none; color: rgb(243, 156, 18); font-weight: 600; background: rgb(255, 255, 255);"
+       onclick="swal('Good job!', 'You clicked the button!', 'success')"
+    >Read more</a>
 </div>