+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog Laravel Currency
 
+## 4.0.0 - 2020-11-19
+
+- Support Laravel 8
+- Rename Account => Profile
+- Upd config
+- Upd auth forms (support laravel/fortify)
+
+## 3.0.0 - 2020-06-10
+
+- Refactor scripts
+- Refactor fields
+- Support Laravel 7
+
 ## 2.0.0 - 2020-02-02
 
 - Release v2